requirements.txt
streamlit
python-pptx
sentence-transformers
faiss-cpu
openai
pyttsx3
google-generativeai>=0.1.3